requirements.txt
python-telegram-bot>=20.3
SQLAlchemy>=1.4.0
python-dotenv>=0.19.0
Pillow>=9.0.0

